requirements.txt
fastapi
uvicorn
pymongo
PyJWT
python-dotenv
pdfplumber
python-docx
spacy
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl